Corrections after comments from CADC on pull request
