--- caom2-repo/src/main/java/ca/nrc/cadc/caom2/repo/client/DeletionState.java
@@ -1,5 +1,0 @@
-package ca.nrc.cadc.caom2.repo.client;
-
-public class DeletionState {
-
-}
